- use survey language picker in survey props - move files for survey props editor into folder
