+++ package.json
@@ -47,14 +47,14 @@
 		"@types/react": "18.3.3",
 		"@types/react-dom": "18.3.0",
 		"case-editor-tools": "^1.6.8",
-		"class-variance-authority": "^0.7.0",
+		"class-variance-authority": "^0.7.1",
 		"clsx": "^2.1.1",
 		"cmdk": "1.0.4",
 		"date-fns": "^3.6.0",
 		"framer-motion": "^11.3.28",
 		"fuse.js": "^7.0.0",
 		"js-yaml": "^4.1.0",
-		"lucide-react": "^0.471.1",
+		"lucide-react": "^0.472.0",
 		"monaco-editor": "^0.49.0",
 		"next": "14.2.23",
 		"next-auth": "5.0.0-beta.22",
